Remove sized requirement from Arc::from_raw
`ptr.byte_sub` is stable since 1.75.0.

The tricky part here is figuring out offset of `ArcInner.data` for
DST.  This PR assumes that raw pointer is coming from existing `Arc`
so it is safe to convert to reference. I'm not 100% sure this is
always correct.

Correct version would require `Layout::for_value_raw` (or
`mem::align_of_val_raw`) which is not stable.

This can be used to convert `Arc<String>` to `Arc<dyn Debug>`.