Derive Copy and Clone HeaderSlice and HeaderWithLength
We (apollo-compiler) have an AST where most nodes are wrapped in,
approximately:

```rust
struct Node<T>(triomphe::Arc<(SourceSpan, T)>);
```

`Arc` allows subtrees to be shared without deep cloning,
and `Arc::make_mut` is used extensively for mutability
with clone-on-write semantics.

Now I would like to also support `Node<str>` and rely on
`Arc::from_header_and_str` for the unsafe bits.
This requires using `HeaderSlice` which with triomphe v0.1.12
breaks `Arc::make_mut` because it does not implement `Clone`.
This PR fixes that.

`Clone` cannot work for `T: !Sized`
which is the main use case for `HeaderSlice` but that’s ok,
`derive` automatically insert the appropriate `where` bound.
This is still useful in order to have the same `Node<T>` use `HeaderSlice`
to support both `T: Sized` and `T: !Sized`.
(The names of `HeaderSlice` and its `slice` field are weird when `T: Sized`
but that weirdness is contained in the module defining `struct Node`.)

While I don’t have a use for this today this PR also derives `Copy`,
and gives `HeaderWithLength` the same treatment.