Add weaker UnwindSafe impls for shared Arc pointers
Following the example of the standard library, it is ok for shared
pointer types to be UnwindSafe when the pointee is only RefUnwindSafe.
This is because UnwindSafe's purpose is to warn against the possibility
of broken invariants being witnessed through the type, which is not
possible with shared pointer types because they don't allow mutation of
the pointee.